requirements.txt
torch==1.10.0+cu113
torchvision==0.11.1+cu113
torchaudio==0.10.0+cu113
PySoundFile==0.9.0
transformers
datasets
sacrebleu
accelerate
tqdm
sentencepiece
pydub
googletrans
speechrecognition
google_cloud_trans